requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==3.0.1
attrs==23.2.0
bcrypt==4.2.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==43.0.0
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==2.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
multidict==6.0.5
openai==1.35.13
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.23.0
rsa==4.9
saucenao-api==2.4.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.13
yarl==1.9.4
